pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry.masonry.api"]

[tool.isort]
profile = "black"
py_version = 39

[tool.mypy]
python_version = "3.9"

[tool.poetry]
authors = [
  "Jérome Eertmans <jeertmans@icloud.com>",
  "Gilles Poncelet <gillesponcelet98@gmail.com>",
  "Louis Navarre <louisnavarre@hotmail.com>",
  "Laurent Dubois <laurent.dubois@uclouvain.be>"
]
description = "ADE Scheduler: a scheduling tool made for humans"
license = "MIT"
name = "ADE Scheduler"
packages = [
  {include = "backend"},
  {include = "cli"},
  {include = "views"}
]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "1.7.4"
authlib = "0.15.5"
bcrypt = "3.2.0"
bleach = "4.1.0"
click = "8.0.3"
cryptography = "3.4.7"
flask = "2.0.2"
flask-babel = "2.0.0"
flask-compress = "1.10.1"
flask-jsglue = "0.3.1"
flask-limiter = ">=2,<3"
flask-login = "0.5.0"
flask-mail = "0.9.1"
flask-migrate = "3.1.0"
flask-session = "0.4.0"
flask-sqlalchemy = "2.5.1"
flask-talisman = ">=0.3.2,<1.0"
geopy = "2.2.0"
ics = "0.7"
jinja2 = "3.0.2"
jsmin = "3.0.1"
lxml = "4.6.5"
numpy = "1.22.1"
pandas = "1.4.0"
plotly = "5.5.0"
psycopg2-binary = "^2.9.9"
pyopenssl = "21.0.0"
python = "^3.9"
python-dotenv = "0.19.2"
pytz = "2021.3"
redis = "3.5.3"
requests = "2.26.0"
sentry-sdk = {version = ">=1.0.0,<2.0.0", extras = ["flask"]}
sqlalchemy = "1.4.22"
tatsu = "5.6.1"
telegram = "0.0.1"
uritools = ">=1.0.1"
user-agents = "2.2.0"
werkzeug = "2.0.2"

[tool.poetry.group.bot.dependencies]
python-dotenv = ">=0.13.0"
python-telegram-bot = "13.10"

[tool.poetry.group.dev.dependencies]
autoflake = ">=1.3.1"
black = "21.12b0"
coverage = ">=5.3,<8"
docker-services-cli = ">=0.4.0"
isort = "^5.13.2"
mock = "^5.1.0"
mypy = "0.931"
pre-commit = "2.17.0"
pydocstyle = "^6.3.0"
pytest-cov = ">=3.0.0"
pytest-flask = "1.2.0"
pytest-isort = ">=3.0.0"
pytest-pycodestyle = ">=2.2.0"
pytest-pydocstyle = ">=2.2.3"

[tool.poetry.group.docs.dependencies]
furo = "^2022.9.29"
myst-parser = "^0.18.1"
sphinx = ">=4.4.0"
sphinx-click = "3.0.2"
sphinx-copybutton = "^0.5.1"
sphinxext-opengraph = "^0.7.5"

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--pycodestyle",
  "--pydocstyle",
  "--isort",
  "--doctest-glob=*.rst",
  "--doctest-modules",
  "--cov-report=term-missing",
  "--cov=backend",
  "--ignore=setup.py",
  "--ignore=docs/source/conf.py"
]
doctest_optionflags = ["NORMALIZE_WHITESPACE"]
testpaths = [
  "docs",
  "tests"
]

[tool.ruff]
line-length = 80
target-version = "py39"
